+++ demo-service/demo-service-server/src/main/java/com/zhw/ms/demo/service/DemoService.java
@@ -15,7 +15,7 @@ public class DemoService {
     private AdminMapper adminMapper;
 
     @SlaverDataSource
-    public Admin getAdmin(Integer id) {
+    public Admin getAdmin(Long id) {
         return adminMapper.selectByPrimaryKey(id);
     }
